Move more google related code to a common api
